2 create page listado partes

+++ client/src/app/features/component/crear-parte/crear-parte.ts
@@ -132,10 +132,13 @@ export class CrearParte implements OnInit {
   }
 
   borrarParte() {
-    this.parteService.deleteParte(Number(this.idUrl)).subscribe({
-      next: () => console.log("Parte borrado.")
-    });
-    this.router.navigate(['/listado-partes'])
+    const confirmar = confirm("¿Seguro que quieres borrar el parte?")
+    if (confirmar) {
+      this.parteService.deleteParte(Number(this.idUrl)).subscribe({
+        next: () => console.log("Parte borrado.")
+      });
+      this.router.navigate(['/listado-partes'])
+    }
   }
 
   //Ayuda Gemini
